# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Win Bingo API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Wallet {
  id: number;
  balance: number;
  currency: string;
}

export interface TopUpInput {
  amount: number;
}

export interface SetNameInput {
  name: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  credit: "credit",
  debit: "debit",
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  description: string;
  createdAt: string;
}

export type GameStateStatus =
  (typeof GameStateStatus)[keyof typeof GameStateStatus];

export const GameStateStatus = {
  lobby: "lobby",
  playing: "playing",
  finished: "finished",
} as const;

export interface GameState {
  id: number;
  status: GameStateStatus;
  calledNumbers: number[];
  nextCallIn: number;
  playerCount: number;
  prizePool: number;
  entryFee: number;
  /** @nullable */
  lastCalledNumber?: number | null;
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerName?: string | null;
}

export interface JoinResult {
  success: boolean;
  message: string;
  wallet: Wallet;
}

export interface BingoCard {
  gameId: number;
  numbers: number[];
  markedNumbers: number[];
}

export interface LeaderboardEntry {
  rank: number;
  playerName: string;
  wins: number;
  totalEarnings: number;
}