# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Win Bingo API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/topup:
    post:
      operationId: topUpWallet
      tags: [wallet]
      summary: Top up wallet balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopUpInput"
      responses:
        "200":
          description: Updated wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallet/setname:
    post:
      operationId: setPlayerName
      tags: [wallet]
      summary: Update the player's display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetNameInput"
      responses:
        "200":
          description: Updated wallet info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
...
components:
  schemas:
...
    Wallet:
      type: object
      required: [id, balance, currency]
      properties:
        id:
          type: integer
        balance:
          type: number
        currency:
          type: string
    TopUpInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
    SetNameInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
...